+++ R/tm_a_regression.R
@@ -13,6 +13,14 @@ NULL
 #' @param response (\code{data_extract_spec} or \code{list} of multiple \code{data_extract_spec})
 #'  Response variables from an incoming dataset with filtering and selecting.
 #' @param default_plot_type optional, (\code{numeric}) Defaults to Response vs Regressor.
+#' @param alpha optional, (\code{numeric}) If scalar then the plot points will have a fixed opacity. If a
+#'   slider should be presented to adjust the plot point opacity dynamically then it can be a vector of
+#'   length three with \code{c(value, min, max)}.
+#' @param size optional, (\code{numeric}) If scalar then the plot point sizes will have a fixed size
+#'   If a slider should be presented to adjust the plot point sizes dynamically then it can be a
+#'   vector of length three with \code{c(value, min, max)}.
+#' @param ggtheme optional, (\code{character}) \code{ggplot} Theme to be used by default.
+#'   All themes can be chosen by the user. Defaults to \code{grey}.
 #' \itemize{
 #'  \item{1 }{Response vs Regressor}
 #'  \item{2 }{Residuals vs Fitted}
@@ -73,10 +81,23 @@ tm_a_regression <- function(label = "Regression Analysis",
                             regressor,
                             response,
                             plot_height = c(600, 200, 2000),
+                            alpha = c(1, 0, 1),
+                            size = c(2, 1, 8),
+                            ggtheme = c(
+                              "grey",
+                              "gray",
+                              "bw",
+                              "linedraw",
+                              "light",
+                              "dark",
+                              "minimal",
+                              "classic",
+                              "void",
+                              "test"
+                            ),
                             pre_output = NULL,
                             post_output = NULL,
-                            default_plot_type = 1
-                            ) {
+                            default_plot_type = 1) {
   if (!is_class_list("data_extract_spec")(regressor)) {
     regressor <- list(regressor)
   }
@@ -87,13 +108,19 @@ tm_a_regression <- function(label = "Regression Analysis",
   stopifnot(is_character_single(label))
   stopifnot(is_class_list("data_extract_spec")(response))
   stop_if_not(list(
-    all(vapply(response, function(x) !isTRUE(x$select$multiple), logical(1))),
-    "Response variable should not allow multiple selection"))
+    all(vapply(response, function(x) {
+      !isTRUE(x$select$multiple)
+    }, logical(1))),
+    "Response variable should not allow multiple selection"
+  ))
   stopifnot(is_class_list("data_extract_spec")(regressor))
-  stopifnot(is_numeric_vector(plot_height) && length(plot_height) == 3)
-  stopifnot(plot_height[1] >= plot_height[2] && plot_height[1] <= plot_height[3])
+  stopifnot(is_numeric_vector(plot_height) &&
+    length(plot_height) == 3)
+  stopifnot(plot_height[1] >= plot_height[2] &&
+    plot_height[1] <= plot_height[3])
   # No check necessary for regressor and response, as checked in data_extract_input
-
+  ggtheme <- match.arg(ggtheme)
+  stopifnot(is_character_single(ggtheme))
   # Send ui args
   args <- as.list(environment())
 
@@ -112,7 +139,6 @@ tm_a_regression <- function(label = "Regression Analysis",
   )
 }
 
-
 ui_a_regression <- function(id, ...) {
   ns <- NS(id)
   args <- list(...)
@@ -120,17 +146,18 @@ ui_a_regression <- function(id, ...) {
   plot_choices <- c(
     "Response vs Regressor",
     "Residuals vs Fitted",
-    "Normal Q-Q", "Scale-Location", "Cook's distance", "Residuals vs Leverage",
+    "Normal Q-Q",
+    "Scale-Location",
+    "Cook's distance",
+    "Residuals vs Leverage",
     "Cook's dist vs Leverage h[ii]/(1 - h[ii])"
   )
 
   standard_layout(
-    output = white_small_well(
-      tags$div(
-        plot_with_settings_ui(id = ns("myplot"), height = args$plot_height),
-        tags$div(verbatimTextOutput(ns("text")))
-      )
-    ),
+    output = white_small_well(tags$div(
+      plot_with_settings_ui(id = ns("myplot"), height = args$plot_height),
+      tags$div(verbatimTextOutput(ns("text")))
+    )),
     encoding = div(
       tags$label("Encodings", class = "text-primary"),
       datanames_input(args[c("response", "regressor")]),
@@ -149,6 +176,31 @@ ui_a_regression <- function(id, ...) {
         label = "Plot type:",
         choices = plot_choices,
         selected = plot_choices[args$default_plot_type]
+      ),
+      panel_group(
+        panel_item(
+          title = "Plot settings",
+          optionalSliderInputValMinMax(ns("alpha"), "Opacity:", args$alpha, ticks = FALSE),
+          optionalSliderInputValMinMax(ns("size"), "Points size:", args$size, ticks = FALSE),
+          optionalSelectInput(
+            inputId = ns("ggtheme"),
+            label = "Theme (by ggplot):",
+            choices = c(
+              "grey",
+              "gray",
+              "bw",
+              "linedraw",
+              "light",
+              "dark",
+              "minimal",
+              "classic",
+              "void",
+              "test"
+            ),
+            selected = args$ggtheme,
+            multiple = FALSE
+          )
+        )
       )
     ),
     forms = get_rcode_ui(ns("rcode")),
@@ -158,11 +210,17 @@ ui_a_regression <- function(id, ...) {
 }
 
 
-#' @importFrom graphics plot abline
 #' @importFrom magrittr %>%
 #' @importFrom methods is substituteDirect
 #' @importFrom stats as.formula
-srv_a_regression <- function(input, output, session, datasets, response, regressor, plot_height) {
+#' @importFrom stats lowess
+srv_a_regression <- function(input,
+                             output,
+                             session,
+                             datasets,
+                             response,
+                             regressor,
+                             plot_height) {
   init_chunks()
 
   merged_data <- data_merge_module(
@@ -173,21 +231,35 @@ srv_a_regression <- function(input, output, session, datasets, response, regress
 
   # sets chunk object and populates it with data merge call and fit expression
   fit <- reactive({
-    chunks_reset()
-    chunks_push_data_merge(merged_data())
+    chunks_stack <- chunks$new()
 
-    ANL <- chunks_get_var("ANL") # nolint
+    chunks_reset(chunks = chunks_stack)
+
+    chunks_push_chunks(merged_data()$chunks, chunks = chunks_stack)
+
+    ANL <- chunks_get_var("ANL", chunks = chunks_stack) # nolint
     validate_has_data(ANL, 10)
 
     response_var <- as.vector(merged_data()$columns_source$response)
     regressor_var <- as.vector(merged_data()$columns_source$regressor)
 
     # validation
-    validate(need(length(regressor_var) > 0, "At least one regressor should be selected."))
-    validate(need(length(response_var) == 1, "Response variable should be of length one."))
+    validate(need(
+      length(regressor_var) > 0,
+      "At least one regressor should be selected."
+    ))
+    validate(need(
+      length(response_var) == 1,
+      "Response variable should be of length one."
+    ))
     validate(need(is.numeric(ANL[response_var][[1]]), "Response variable should be numeric."))
     if (input$plot_type == "Response vs Regressor") {
-      validate(need(length(regressor_var) == 1, "Response vs Regressor is only provided for exactly one regressor"))
+      validate(
+        need(
+          length(regressor_var) == 1,
+          "Response vs Regressor is only provided for exactly one regressor"
+        )
+      )
     }
 
     validate_has_data(ANL[, c(response_var, regressor_var)], 10, complete = TRUE)
@@ -204,56 +276,242 @@ srv_a_regression <- function(input, output, session, datasets, response, regress
     )
 
     chunks_push(
-      expression = quote(fit <- lm(form, data = ANL)) %>% substituteDirect(list(form = form))
+      id = "form",
+      expression = bquote(form <- .(form)),
+      chunks = chunks_stack
     )
-    chunks_safe_eval()
-    return(form)
+
+    chunks_push(
+      id = "fit_lm",
+      expression = quote(fit <- lm(form, data = ANL)) %>% substituteDirect(list(form = form)),
+      chunks = chunks_stack
+    )
+
+    chunks_safe_eval(chunks = chunks_stack)
+
+    chunks_stack
   })
 
   plot_r <- reactive({
-    fit()
+    alpha <- input$alpha # nolint
+    size <- input$size # nolint
+    ggtheme <- input$ggtheme # nolint
+    input_type <- input$plot_type
 
-    if (input$plot_type == "Response vs Regressor") {
+    chunks_reset()
+    chunks_push_chunks(fit())
+
+    validate(need(!is.null(ggtheme), "Please select a theme."))
+
+    plot_type_0 <- function() {
       fit <- chunks_get_var("fit") # chunk already evaluated
-      response_var <- as.vector(merged_data()$columns_source$response)
-      regressor_var <- as.vector(merged_data()$columns_source$regressor)
+      response_var <- as.vector(merged_data()$columns_source$response) # nolint
+      regressor_var <- as.vector(merged_data()$columns_source$regressor) # nolint
       ANL <- chunks_get_var("ANL") # nolint
 
-      if (ncol(fit$model) > 1) {
-        stopifnot(ncol(fit$model) == 2)
-        chunks_push(bquote({
-          plot(
+      stopifnot(ncol(fit$model) == 2)
+
+      if (!is.factor(ANL[[regressor_var]])) {
+        shinyjs::show("size")
+        shinyjs::show("alpha")
+        chunks_push(id = "plot_0_a", expression = bquote({
+          gg <- ggplot(
             fit$model[, 2:1],
-            main = "Response vs Regressor",
-            xlab = .(varname_w_label(regressor_var, ANL)),
-            ylab = .(varname_w_label(response_var, ANL)))
+            aes_string(.(regressor_var), .(response_var))
+          ) +
+            geom_point(size = .(size), alpha = .(alpha)) +
+            stat_smooth(
+              method = "lm",
+              formula = y ~ x,
+              se = FALSE
+            )
         }))
-        if (is.numeric(ANL[[regressor_var]])) {
-          chunks_push(quote(abline(fit, col = "red", lwd = 2L)))
-        }
       } else {
-        ANL <- chunks_get_var("ANL") # nolint
-        chunks_push(quote({
-          plot_data <- data.frame(fit$model[, 1], fit$model[, 1])
-          names(plot_data) <- rep(names(fit$model), 2)
-          plot <- plot(plot_data, main = "Response vs Regressor")
-          abline(ANL)
+        shinyjs::hide("size")
+        shinyjs::hide("alpha")
+        chunks_push(id = "plot_0_a", expression = bquote({
+          gg <- ggplot(
+            fit$model[, 2:1],
+            aes_string(.(regressor_var), .(response_var))
+          ) + geom_boxplot()
         }))
       }
+
+      chunks_push(id = "plot_0_b", expression = bquote({
+        g_final <- gg +
+          xlab(.(varname_w_label(regressor_var, ANL))) +
+          ylab(.(varname_w_label(response_var, ANL))) +
+          ggtitle("Response vs Regressor") +
+          .(call(paste0("theme_", ggtheme)))
+        print(g_final)
+      }))
+    }
+
+    plot_base <- function() {
+      chunks_push(id = "plot_0_c", expression = bquote({
+        class(fit$residuals) <- NULL
+
+        data <- fortify(fit)
+
+        smooth <- function(x, y) {
+          as.data.frame(stats::lowess(x, y, f = 2 / 3, iter = 3))
+        }
+
+        smoothy_aes <- ggplot2::aes_string(x = "x", y = "y")
+
+        reg_form <- deparse(fit$call[[2]])
+      }))
+    }
+
+    plot_type_1 <- function() {
+      shinyjs::show("size")
+      shinyjs::show("alpha")
+      chunks_push(id = "plot_1", expression = bquote({
+        smoothy <- smooth(data$.fitted, data$.resid)
+        g_base <- ggplot(data = data, aes(.fitted, .resid)) +
+          geom_point(size = .(size), alpha = .(alpha)) +
+          geom_hline(
+            yintercept = 0,
+            linetype = "dashed",
+            size = 1
+          ) +
+          geom_line(data = smoothy, mapping = smoothy_aes)
+        g_final <- g_base + labs(
+          x = paste0("Fitted values\nlm(", reg_form, ")"),
+          y = "Residuals",
+          title = .(input_type)
+        ) + .(call(paste0("theme_", ggtheme)))
+        print(g_final)
+      }))
+    }
+
+    plot_type_2 <- function() {
+      shinyjs::show("size")
+      shinyjs::show("alpha")
+      chunks_push(id = "plot_2", expression = bquote({
+        g_base <- ggplot(data = data) +
+          stat_qq(aes(sample = .stdresid), size = .(size), alpha = .(alpha)) +
+          geom_abline(linetype = "dashed")
+        g_final <- g_base + labs(
+          x = paste0("Theoretical Quantiles\nlm(", reg_form, ")"),
+          y = "Standardized residuals",
+          title = .(input_type)
+        ) +
+          .(call(paste0("theme_", ggtheme)))
+        print(g_final)
+      }))
+    }
+
+    plot_type_3 <- function() {
+      shinyjs::show("size")
+      shinyjs::show("alpha")
+      chunks_push(id = "plot_3", expression = bquote({
+        smoothy <- smooth(data$.fitted, sqrt(abs(data$.stdresid)))
+        g_base <- ggplot(data = data, aes(.fitted, sqrt(abs(.stdresid)))) +
+          geom_point(size = .(size), alpha = .(alpha)) +
+          geom_line(data = smoothy, mapping = smoothy_aes)
+        g_final <- g_base + labs(
+          x = paste0("Fitted values\nlm(", reg_form, ")"),
+          y = expression(sqrt(abs(`Standardized residuals`))),
+          title = .(input_type)
+        ) +
+          .(call(paste0("theme_", ggtheme)))
+        print(g_final)
+      }))
+    }
+
+    plot_type_4 <- function() {
+      shinyjs::hide("size")
+      shinyjs::show("alpha")
+      chunks_push(id = "plot_4", expression = bquote({
+        g_base <- ggplot(data = data, aes(seq_along(.cooksd), .cooksd)) +
+          geom_col(alpha = .(alpha))
+        g_final <- g_base + labs(
+          x = paste0("Obs. number\nlm(", reg_form, ")"),
+          y = "Cook's distance",
+          title = .(input_type)
+        ) +
+          .(call(paste0("theme_", ggtheme)))
+        print(g_final)
+      }))
+    }
+
+
+    plot_type_5 <- function() {
+      shinyjs::show("size")
+      shinyjs::show("alpha")
+      chunks_push(id = "plot_5", expression = bquote({
+        smoothy <- smooth(data$.hat, data$.stdresid)
+        g_base <- ggplot(data = data, aes(.hat, .stdresid)) +
+          geom_vline(
+            size = 1,
+            colour = "black",
+            linetype = "dashed",
+            xintercept = 0
+          ) +
+          geom_hline(
+            size = 1,
+            colour = "black",
+            linetype = "dashed",
+            yintercept = 0
+          ) +
+          geom_point(size = .(size), alpha = .(alpha)) +
+          geom_line(data = smoothy, mapping = smoothy_aes)
+        g_final <- g_base + labs(
+          x = paste0("Standardized residuals\nlm(", reg_form, ")"),
+          y = "Leverage",
+          title = .(input_type)
+        ) +
+          .(call(paste0("theme_", ggtheme)))
+        print(g_final)
+      }))
+    }
+
+
+    plot_type_6 <- function() {
+      shinyjs::show("size")
+      shinyjs::show("alpha")
+      chunks_push(id = "plot_6", expression = bquote({
+        smoothy <- smooth(data$.hat, data$.cooksd)
+        g_base <- ggplot(data = data, aes(.hat, .cooksd)) +
+          geom_vline(xintercept = 0, colour = NA) +
+          geom_abline(
+            slope = seq(0, 3, by = 0.5),
+            colour = "black",
+            linetype = "dashed",
+            size = 1
+          ) +
+          geom_line(data = smoothy, mapping = smoothy_aes) +
+          geom_point(size = .(size), alpha = .(alpha))
+        g_final <- g_base + labs(
+          x = paste0("Leverage\nlm(", reg_form, ")"),
+          y = "Cooks's distance",
+          title = .(input_type)
+        ) +
+          .(call(paste0("theme_", ggtheme)))
+        print(g_final)
+      }))
+    }
+
+    if (input_type == "Response vs Regressor") {
+      plot_type_0()
     } else {
-      i <- which(input$plot_type == c(
-        "Residuals vs Fitted",
-        "Normal Q-Q",
-        "Scale-Location",
-        "Cook's distance",
-        "Residuals vs Leverage",
-        "Cook's dist vs Leverage h[ii]/(1 - h[ii])"
-      ))
-      chunks_push(bquote(plot(fit, which = .(i), id.n = NULL)))
+      plot_base()
+      switch(input_type,
+        "Residuals vs Fitted"  = plot_type_1(),
+        "Normal Q-Q" = plot_type_2(),
+        "Scale-Location" =  plot_type_3(),
+        "Cook's distance" =  plot_type_4(),
+        "Residuals vs Leverage" =  plot_type_5(),
+        "Cook's dist vs Leverage h[ii]/(1 - h[ii])" =  plot_type_6()
+      )
     }
 
     chunks_safe_eval()
   })
+
+
+  code_header <- reactive(chunks_get_var("form", chunks = fit()))
 
   # Insert the plot into a plot_with_settings module from teal.devel
   callModule(
@@ -264,7 +522,8 @@ srv_a_regression <- function(input, output, session, datasets, response, regress
   )
 
   output$text <- renderPrint({
-    fit()
+    chunks_reset()
+    chunks_push_chunks(fit())
     chunks_push(expression = quote(summary(fit)), id = "summary")
     chunks_safe_eval()
   })
@@ -277,7 +536,7 @@ srv_a_regression <- function(input, output, session, datasets, response, regress
     modal_title = "R code for the regression plot",
     code_header = paste0(
       "Regression plot of ",
-      format(fit())
+      format(code_header())
     )
   )
 }